refactor: moved database interface to model
